+++ BambooTracker/gui/order_list_editor/order_list_panel.hpp
@@ -83,6 +83,7 @@ public slots:
 	void onCloneOrderPressed();
 	void onFollowModeChanged();
 	void onStoppedPlaySong();
+	void onGoOrderRequested(bool toNext);
 
 signals:
 	void hScrollBarChangeRequested(int num);